Create RequireDepartmentNameTagKeyWithSpecificValue.json
Enforce the use of a DepartmentName tag key with a list of specific departments at resource group scope.  This is handy when paired with policies that inherit tags from RG scope - it ensures that an RG owned or used by a single department not only is named for the department but is against a standardized list.  This might be cost centers in other organizations.